requirements.txt
Flask
Flask-RESTful
Flask-SQLAlchemy
flask-apispec>=0.11,<1.0
marshmallow>=3.14,<4
# psycopg[binary]==3.*